Move runner types under version namespaces
