+++ src/extension/grpc/runner/v2alpha1.ts
@@ -1,0 +1,1 @@
+export * from '@buf/stateful_runme.community_timostamm-protobuf-ts/runme/runner/v2alpha1/runner_pb'
