+++ tests/extension/executors/task.test.ts
@@ -24,7 +24,7 @@ vi.mock('node:fs/promises', () => ({
   },
 }))
 
-vi.mock('../../../src/extension/grpc/runnerTypes', () => ({
+vi.mock('../../../src/extension/grpc/runner/v1', () => ({
   CommandMode: {
     INLINE_SHELL: 1,
     TEMP_FILE: 2,
