+++ tests/extension/messages/platformApiRequest/saveCellExecution.test.ts
@@ -13,7 +13,7 @@ import { getCellById } from '../../../../src/extension/cell'
 
 vi.mock('vscode-telemetry')
 vi.mock('../../../src/extension/runner', () => ({}))
-vi.mock('../../../src/extension/grpc/runnerTypes', () => ({}))
+vi.mock('../../../src/extension/grpc/runner/v1', () => ({}))
 
 vi.mock('vscode', async () => {
   const mocked = (await vi.importActual('../../../../__mocks__/vscode')) as any
